requirements.txt
# Auto Generated
Pillow
bs4
gig-nuuuwan
pytesseract
tesseract
tesseract-ocr
tweepy
utils-nuuuwan
